created src/crop.py and expanded src/path.py
