+++ src/enums.py
@@ -1,16 +1,18 @@
-from enum import Enum
+#==============================================================================
+# enums.py 
+#     Enums live here. 
+#==============================================================================
 
+from enum import Enum
 
 class Registration(str, Enum):
     THREE = "3"
     FOUR = "4"
 
-
 # Paper orientation: portrait keeps cards upright, landscape rotates them 90 degrees.
 class Orientation(str, Enum):
     PORTRAIT = "portrait"
     LANDSCAPE = "landscape"
-
 
 class OrientationMode(str, Enum):
     """CLI orientation selection. OPTIMIZE tries both and picks the best."""
@@ -18,16 +20,13 @@ class OrientationMode(str, Enum):
     LANDSCAPE = Orientation.LANDSCAPE.value
     PORTRAIT = Orientation.PORTRAIT.value
 
-
 class Variant(str, Enum):
     DEFAULT = "default"
     BORDERLESS = "borderless"
 
-
 class Unit(str, Enum):
     MM = "mm"
     IN = "in"
-
 
 class FitMode(str, Enum):
     STRETCH = "stretch"
